DPSWarr: Removed unnecessary float casts on Math.Max and Math.Min.  Also removed string calculations when needsDisplay is false.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C37638

+++ Rawr.DPSWarr/CalculationOptionsDPSWarr.cs
@@ -109,8 +109,8 @@ namespace Rawr.DPSWarr {
 		// Latency
 		public float Lag = 179f;
 		public float React = 220f;
-        public float Latency { get { return (float)Math.Max(0f, Lag / 1000f); } }
-        public float AllowedReact { get { return (float)Math.Max(0f, (React - 250f) / 1000f); } }
+        public float Latency { get { return Math.Max(0f, Lag / 1000f); } }
+        public float AllowedReact { get { return Math.Max(0f, (React - 250f) / 1000f); } }
         public float FullLatency { get { return AllowedReact + Latency; } }
 		//
 		public WarriorTalents talents = null;
